prefactoring: merge two test modules into one which are logically the same thing
